Raise what a zindex names, whatever its sign
`visit` partitions on `if (item.zindex)` and only then sorts:

    function compare(a, b) { return a.zindex - b.zindex || a.index - b.index; }
    ...
    for (i=0; i<n; ++i) if (!items[i].zindex) visitor(items[i]);
    items = zitems;

So a **negative** zindex does not sink a mark or an axis below its neighbours:
it raises it above all of them and puts it at the bottom of the raised ones.
Axes were partitioned here by sign, so `zindex: -2` sank; the raised ones were
then drawn in declaration order rather than by upstream's comparator; and every
zindex was truncated to an integer, so `zindex: 0.5` was read as none at all.
`paintOrder` already had the rule right for marks and only needed the number.

`zindex-sign` is the fixture: a bottom axis at -2, a pair of bottom axes whose
declaration order the fractional one reverses, a grid axis at 0.5, a rect at -1
and a rule at 0.25. Four arms mutation-checked.

API: `zindex` is a `Double` on the mark, axis, legend and title specs, where it
was an `Int`.

+++ vega-model/src/main/kotlin/dev/aster/vega/model/spec/VegaSpec.kt
@@ -784,7 +784,7 @@ public data class TitleSpec(
   val dx: NumberValue? = null,
   val dy: NumberValue? = null,
   val subtitleFontSize: NumberValue? = null,
-  val zindex: Int = 0,
+  val zindex: Double = 0.0,
   /**
    * `encode`, keyed by the part it addresses: `group`, `title` or `subtitle`.
    *
@@ -936,7 +936,7 @@ public data class AxisSpec(
   val titleAngle: NumberValue? = null,
   val titleAlign: String? = null,
   val titleBaseline: String? = null,
-  val zindex: Int = 0,
+  val zindex: Double = 0.0,
   /**
    * Explicit tick values, replacing the ones the scale would generate.
    *
@@ -1437,7 +1437,7 @@ public data class LegendSpec(
   val description: String? = null,
   val backgroundStrokeWidth: Double? = null,
   val backgroundStrokeDash: List<Double>? = null,
-  val zindex: Int = 0,
+  val zindex: Double = 0.0,
   /**
    * Appearance of the three parts, read the same way an axis reads its own.
    *
@@ -1976,7 +1976,7 @@ public data class MarkSpec(
   val layout: LayoutSpec? = null,
   /** A title for this group, which is how a trellis header labels its row or column. */
   val title: TitleSpec? = null,
-  val zindex: Int = 0,
+  val zindex: Double = 0.0,
   val interactive: Boolean = true,
   /**
    * `aria: false` hides this mark from a screen reader entirely.
